+++ playwright.config.ts
@@ -44,7 +44,7 @@ export default defineConfig({
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: 'on-first-retry',
     /* Configure the timeout for each test */
-    navigationTimeout: 60000,
+    navigationTimeout: 3000,
   },
 
   /* Configure projects for major browsers */